Broken down the big handleChatMessage
